Revert "Add AggLayer module (#5110)"
This reverts commit d7a16859267e931537805d230b997201893cfc52.
