+++ packages/backend/src/config/features/tvl.ts
@@ -14,7 +14,6 @@ import {
 } from '@l2beat/config'
 import { UnixTime } from '@l2beat/shared-pure'
 
-import { uniq } from 'lodash'
 import { TvlConfig } from '../Config'
 import { FeatureFlags } from '../FeatureFlags'
 import { getChainTvlConfig, getChainsWithTokens } from './chains'
@@ -29,26 +28,13 @@ export function getTvlConfig(
     .concat(bridges.map(bridgeToBackendProject))
     .concat(layer3s.map(layer3ToBackendProject))
 
-  const aggLayerEnabled = flags.isEnabled('tvl', 'aggLayer')
-
-  const aggLayerChains = layer2s
-    .filter((c) =>
-      c.config.escrows.some((e) => e.sharedEscrow?.type === 'AggLayer'),
-    )
-    .map((l) => l.id)
-
-  const chainConfigs = uniq(
-    getChainsWithTokens(tokenList, chains).concat(
-      aggLayerEnabled ? aggLayerChains : [],
-    ),
-  ).map((chain) =>
+  const chainConfigs = getChainsWithTokens(tokenList, chains).map((chain) =>
     getChainTvlConfig(flags.isEnabled('tvl', chain), env, chain, {
       minTimestamp: minTimestampOverride,
     }),
   )
 
   return {
-    aggLayerEnabled,
     amounts: getTvlAmountsConfig(projects),
     prices: getTvlPricesConfig(minTimestampOverride),
     chains: chainConfigs,
