+++ packages/backend/src/modules/tvl/api/services/TvlService.ts
@@ -363,17 +363,10 @@ export class TvlService {
         priceUsd: price,
         decimals: config.decimals,
       })
-      const address =
-        config.type === 'aggLayerL2Token'
-          ? config.l1Address.toString()
-          : config.type === 'aggLayerNativeEtherPreminted' ||
-              config.type === 'aggLayerNativeEtherWrapped'
-            ? 'native'
-            : config.address.toString()
 
       breakdown[config.source].push({
         assetId: priceConfig.assetId,
-        address,
+        address: config.address.toString(),
         chain: config.chain,
         chainId: this.$.chainConverter.toChainId(config.chain),
         source: config.source,
