+++ packages/backend/src/modules/tvl/services/AggLayerService.ts
@@ -6,7 +6,6 @@ import {
   Bytes,
   EthereumAddress,
   UnixTime,
-  notUndefined,
 } from '@l2beat/shared-pure'
 import { BigNumber, utils } from 'ethers'
 import { MulticallClient } from '../../../peripherals/multicall/MulticallClient'
@@ -107,13 +106,6 @@ export class AggLayerService {
 
     return responses.map((response, index) => {
       const token = tokens[index]
-      if (response.data.toString() === '0x') {
-        return {
-          configId: token.id,
-          timestamp,
-          amount: 0n,
-        }
-      }
       const [value] = erc20Interface.decodeFunctionResult(
         'totalSupply',
         response.data.toString(),
@@ -145,23 +137,17 @@ export class AggLayerService {
       blockNumber,
     )
 
-    return responses
-      .map((response, index) => {
-        const token = tokens[index]
-        if (response.data.toString() === '0x') {
-          return
-        }
-
-        const [address] = bridgeInterface.decodeFunctionResult(
-          'getTokenWrappedAddress',
-          response.data.toString(),
-        )
-        return {
-          ...token,
-          address,
-        }
-      })
-      .filter(notUndefined)
+    return responses.map((response, index) => {
+      const token = tokens[index]
+      const [address] = bridgeInterface.decodeFunctionResult(
+        'getTokenWrappedAddress',
+        response.data.toString(),
+      )
+      return {
+        ...token,
+        address: address as EthereumAddress,
+      }
+    })
   }
 
   async getNativeEtherPremintedAmount(
