+++ packages/backend/src/modules/tvl/services/ValueService.test.ts
@@ -5,7 +5,6 @@ import {
   AssetId,
   EthereumAddress,
   ProjectId,
-  TotalSupplyEntry,
   UnixTime,
 } from '@l2beat/shared-pure'
 import { expect, mockObject } from 'earl'
@@ -37,7 +36,7 @@ describe(ValueService.name, () => {
     const project: ProjectId = ProjectId('project')
     const source: string = 'chain'
     const addressA = EthereumAddress.random()
-    const CONFIG_A = mockObject<TotalSupplyEntry>({
+    const CONFIG_A = mockObject<AmountConfigEntry>({
       assetId: AssetId.create('chain', addressA),
       sinceTimestamp: UnixTime.ZERO,
       address: addressA,
@@ -49,7 +48,7 @@ describe(ValueService.name, () => {
       category: 'ether',
     })
     const addressB = EthereumAddress.random()
-    const CONFIG_B = mockObject<TotalSupplyEntry>({
+    const CONFIG_B = mockObject<AmountConfigEntry>({
       assetId: AssetId.create('chain', addressB),
       sinceTimestamp: new UnixTime(300),
       address: addressB,
