+++ packages/backend/src/modules/tvl/utils/getValuesConfigHash.ts
@@ -14,7 +14,9 @@ export function getValuesConfigHash(
   const valueIds: string[] = []
 
   for (const amount of amountConfigs) {
-    const price = priceConfigs.find((p) => p.assetId === amount.assetId)
+    const price = priceConfigs.find(
+      (p) => p.address === amount.address && p.chain === amount.chain,
+    )
     assert(price, `Price config not found for ${createAmountId(amount)}`)
     const valueId = createValueId(amount, price)
     valueIds.push(valueId)
