+++ packages/backend/src/modules/tvl/utils/test/mocks.ts
@@ -41,7 +41,6 @@ function amountConfiguration(v: Partial<TotalSupplyEntry>) {
     sinceTimestamp: UnixTime.ZERO,
     includeInTotal: true,
     decimals: DECIMALS,
-    assetId: AssetId.create(v.chain ?? 'chain', v.address),
     symbol: 'SYMBOL',
     isAssociated: false,
     category: 'other',
