+++ packages/config/src/backend/BackendProject.ts
@@ -11,10 +11,7 @@ import {
   TokenBridgedUsing,
   UnixTime,
 } from '@l2beat/shared-pure'
-import {
-  AggLayerEscrow,
-  ScalingProjectEscrow,
-} from '../common/ScalingProjectEscrow'
+import { ScalingProjectEscrow } from '../common/ScalingProjectEscrow'
 import { ScalingProjectTransactionApi } from '../common/ScalingProjectTransactionApi'
 import {
   Bridge,
@@ -50,7 +47,6 @@ export interface BackendProjectEscrow {
   includeInTotal?: boolean
   source?: ScalingProjectEscrow['source']
   bridgedUsing?: TokenBridgedUsing
-  sharedEscrow?: AggLayerEscrow
 }
 
 export function toBackendProject(
@@ -194,7 +190,6 @@ function toProjectEscrow(escrow: ScalingProjectEscrow): BackendProjectEscrow {
     includeInTotal: escrow.includeInTotal,
     source: escrow.source,
     bridgedUsing: escrow.bridgedUsing,
-    sharedEscrow: escrow.sharedEscrow,
   }
 }
