+++ packages/config/src/discovery/ProjectDiscovery.ts
@@ -31,10 +31,7 @@ import {
   ScalingProjectContractSingleAddress,
   ScalingProjectUpgradeability,
 } from '../common/ScalingProjectContracts'
-import {
-  AggLayerEscrow,
-  ScalingProjectEscrow,
-} from '../common/ScalingProjectEscrow'
+import { ScalingProjectEscrow } from '../common/ScalingProjectEscrow'
 import {
   ScalingProjectPermission,
   ScalingProjectPermissionedAccount,
@@ -119,7 +116,6 @@ export class ProjectDiscovery {
     bridgedUsing,
     isHistorical,
     untilTimestamp,
-    sharedEscrow,
   }: {
     address: EthereumAddress
     name?: string
@@ -136,7 +132,6 @@ export class ProjectDiscovery {
     bridgedUsing?: TokenBridgedUsing
     isHistorical?: boolean
     untilTimestamp?: UnixTime
-    sharedEscrow?: AggLayerEscrow
   }): ScalingProjectEscrow {
     const contractRaw = this.getContract(address.toString())
     const timestamp = sinceTimestamp?.toNumber() ?? contractRaw.sinceTimestamp
@@ -170,7 +165,6 @@ export class ProjectDiscovery {
       bridgedUsing,
       isHistorical,
       untilTimestamp,
-      sharedEscrow,
     }
   }
