+++ packages/config/src/projects/layer2s/astarzkevm.ts
@@ -1,4 +1,3 @@
-import { EthereumAddress, UnixTime } from '@l2beat/shared-pure'
 import { NEW_CRYPTOGRAPHY, RISK_VIEW } from '../../common'
 import { ProjectDiscovery } from '../../discovery/ProjectDiscovery'
 import { Badge } from '../badges'
@@ -6,9 +5,6 @@ import { polygonCDKStack } from './templates/polygonCDKStack'
 import { Layer2 } from './types'
 
 const discovery = new ProjectDiscovery('astarzkevm')
-
-const shared = new ProjectDiscovery('shared-polygon-cdk')
-const bridge = shared.getContract('Bridge')
 
 const membersCountDAC = discovery.getContractValue<number>(
   'AstarValidiumDAC',
@@ -96,34 +92,10 @@ export const astarzkevm: Layer2 = polygonCDKStack({
     },
     activityDataSource: 'Blockchain RPC',
   },
-  chainConfig: {
-    name: 'astarzkevm',
-    chainId: 3776,
-    explorerUrl: 'https://astar-zkevm.explorer.startale.com',
-    minTimestampForTvl: new UnixTime(1708632059),
-    multicallContracts: [
-      {
-        address: EthereumAddress('0xcA11bde05977b3631167028862bE2a173976CA11'),
-        batchSize: 150,
-        sinceBlock: 183817,
-        version: '3',
-      },
-    ],
-  },
-  rpcUrl: 'https://rpc.startale.com/astar-zkevm',
+  rpcUrl: 'https://evm.astar.network',
   discovery,
   isForcedBatchDisallowed,
-  nonTemplateEscrows: [
-    shared.getEscrowDetails({
-      address: bridge.address,
-      tokens: '*',
-      sharedEscrow: {
-        type: 'AggLayer',
-        nativeAsset: 'etherPreminted',
-        premintedAmount: '340282366920938463463374607431768211455',
-      },
-    }),
-  ],
+  nonTemplateEscrows: [],
   nonTemplateTechnology: {
     newCryptography: {
       ...NEW_CRYPTOGRAPHY.ZK_BOTH,
