+++ packages/config/src/tvl/ConfigMapping.ts
@@ -5,10 +5,8 @@ import {
   type EthereumAddress,
   type PriceConfigEntry,
   ProjectId,
-  assertUnreachable,
 } from '@l2beat/shared-pure'
 import { groupBy } from 'lodash'
-import { ethereum } from '../chains/ethereum'
 import { createAmountId } from './createAmountId'
 import { createPriceId } from './createPriceId'
 
@@ -44,24 +42,8 @@ export class ConfigMapping {
   getPriceConfigFromAmountConfig(
     amountConfig: AmountConfigEntry,
   ): PriceConfigEntry & { configId: string } {
-    let assetId: AssetId
-    switch (amountConfig.type) {
-      case 'circulatingSupply':
-      case 'preminted':
-      case 'totalSupply':
-      case 'escrow':
-        assetId = amountConfig.assetId
-        break
-      case 'aggLayerL2Token':
-        assetId = AssetId.create(ethereum.name, amountConfig.l1Address)
-        break
-      case 'aggLayerNativeEtherPreminted':
-      case 'aggLayerNativeEtherWrapped':
-        assetId = AssetId.create(ethereum.name, 'native')
-        break
-      default:
-        assertUnreachable(amountConfig)
-    }
+    const assetId = AssetId.create(amountConfig.chain, amountConfig.address)
+
     const priceConfig = this.pricesByAssetId.get(assetId)
 
     assert(
@@ -83,7 +65,9 @@ export class ConfigMapping {
     assert(projectAmounts)
 
     const assetId = AssetId.create(token.chain, token.address)
-    const amountConfigs = projectAmounts.filter((x) => x.assetId === assetId)
+    const amountConfigs = projectAmounts.filter(
+      (x) => AssetId.create(x.chain, x.address) === assetId,
+    )
     assert(
       amountConfigs.every((x) => x.decimals === amountConfigs[0]?.decimals),
       'Decimals mismatch!',
