--- packages/config/src/tvl/aggLayer.ts
@@ -1,6 +1,0 @@
-import { EthereumAddress } from '@l2beat/shared-pure'
-
-// We are assuming that the BRIDGE_ADDRESS is the same on all chains
-export const AGGLAYER_L2BRIDGE_ADDRESS = EthereumAddress(
-  '0x2a3DD3EB832aF982ec71669E178424b10Dca2EDe',
-)
