+++ packages/config/src/tvl/index.ts
@@ -4,4 +4,3 @@ export * from './createPriceId'
 export * from './getTvlAmountsConfig'
 export * from './getTvlPricesConfig'
 export * from './indexerIdentity'
-export * from './aggLayer'
