+++ packages/frontend/src/server/features/scaling/tvl/sync-status/get-amounts-status.ts
@@ -1,5 +1,4 @@
 import { type AmountConfigEntry, type UnixTime } from '@l2beat/shared-pure'
-import { getAggLayerStatus } from './get-agglayer-status'
 import { getCirculatingSupplyStatus } from './get-circulating-supply-status'
 import { getConfigurationsStatus } from './get-configurations-status'
 import { getPremintedStatus } from './get-preminted-status'
@@ -8,11 +7,10 @@ export async function getAmountsStatus(
   entries: (AmountConfigEntry & { configId: string })[],
   targetTimestamp: UnixTime,
 ) {
-  const [configurations, preminted, circulating, aggLayer] = await Promise.all([
+  const [configurations, preminted, circulating] = await Promise.all([
     getConfigurationsStatus(entries, targetTimestamp),
     getPremintedStatus(entries, targetTimestamp),
     getCirculatingSupplyStatus(entries, targetTimestamp),
-    getAggLayerStatus(entries, targetTimestamp),
   ])
 
   return {
@@ -20,13 +18,11 @@ export async function getAmountsStatus(
       ...configurations.lagging,
       ...preminted.lagging,
       ...circulating.lagging,
-      ...aggLayer.lagging,
     ],
     excluded: new Set([
       ...configurations.excluded,
       ...preminted.excluded,
       ...circulating.excluded,
-      ...aggLayer.excluded,
     ]),
   }
 }
