+++ packages/frontend/src/server/features/scaling/tvl/sync-status/to-indexer-entry.ts
@@ -1,0 +1,21 @@
+import { getChainAmountIndexerId, getPriceIndexerId } from '@l2beat/config'
+import {
+  type CoingeckoPriceConfigEntry,
+  type EscrowEntry,
+  type TotalSupplyEntry,
+} from '@l2beat/shared-pure'
+
+export type MultiIndexerEntry =
+  | TotalSupplyEntry
+  | EscrowEntry
+  | CoingeckoPriceConfigEntry
+
+export function toIndexerId(config: MultiIndexerEntry) {
+  switch (config.type) {
+    case 'coingecko':
+      return getPriceIndexerId(config.coingeckoId)
+    case 'escrow':
+    case 'totalSupply':
+      return getChainAmountIndexerId(config.chain)
+  }
+}
