+++ packages/shared-pure/src/types/AmountConfigEntry.ts
@@ -16,9 +16,6 @@ export type AmountConfigEntry =
   | CirculatingSupplyEntry
   | EscrowEntry
   | PremintedEntry
-  | AggLayerL2Token
-  | AggLayerNativeEtherPreminted
-  | AggLayerNativeEtherWrapped
 
 export interface PremintedEntry extends AmountConfigBase {
   type: 'preminted'
@@ -64,18 +61,15 @@ export interface AggLayerL2Token extends AmountConfigBase {
   type: 'aggLayerL2Token'
   l1Address: EthereumAddress
   originNetwork: number
-  escrowAddress: EthereumAddress
 }
 
 export interface AggLayerNativeEtherPreminted extends AmountConfigBase {
   type: 'aggLayerNativeEtherPreminted'
   l2BridgeAddress: EthereumAddress
   premintedAmount: bigint
-  escrowAddress: EthereumAddress
 }
 
 export interface AggLayerNativeEtherWrapped extends AmountConfigBase {
   type: 'aggLayerNativeEtherWrapped'
   wethAddress: EthereumAddress
-  escrowAddress: EthereumAddress
 }
